Highly increasing the performance of Communities3
The parameters are now slightly changed:
Instead of the exact number of inter-community edges, a progrability matrix double[][] p is used to represent the amount of links between communities.
Also, two basic version of the algorithm exist: One uses mor memory, the other one requires more computation time.
It can be changed in the Communities3.LESS_MEMORY flag.

One method is added to the Util class to add a toFolderString method for double[][].