different node orders for the Communities3 network generator
The parameter order is added to the constructor.

order and bidirectional are now part of the configuration parameters.

The parameter "order" can be used to change the position of the nodes in the
main node array of the generated graph object. If it is set to SORTED_ORDER,
the main array contains first all nodes from the first community, then the
nodes from the second community, and so on. In case the flag is true, the
first node is assigned to the first community, the second node to the second
and so on. If the communities do not have the same size, the smaller
communities are simply discarded after they received all their nodes. E.g.,
sizes = {2, 3} would lead to the communities with node indices {0, 2} {1, 3,
4}. sizes = {3, 1, 3} would result in the node assignments {0, 3, 5} {1} {2,
4, 6}. If order is set to RANDOM_ORDER, the nodes are assigned to the
communities at random.

Also, a getter for the last generated community structure is now included.