adds the GreedyBacktrackingExtended routing algorithm
This algorithm work like the GreedyBacktracking routing.
The difference is that a node does not immediately backtrack the message in case it encounters a dead-end.
It first forwards the message once to the next best neighbor (not yet visited and not as close to the destination as the node itself but the neighbor with the shortest distacne remaining).
If this is unsuccessfull as well, the node backtracks the message to the previous node.